+++ src/Suru.Compiler/Parse/AstPrinter.cs
@@ -151,9 +151,9 @@ public static class AstPrinter
 
             case StructLiteralExpression st:
                 sb.AppendLine($"{indent}StructLiteral [{st.Fields.Count} fields]");
-                foreach (var (name, typeAnn, val) in st.Fields)
+                foreach (var (name, val) in st.Fields)
                 {
-                    sb.AppendLine($"{indent}  Field [{name} {typeAnn}]");
+                    sb.AppendLine($"{indent}  Field [{name}]");
                     PrintExpression(sb, indent + "    ", val);
                 }
                 break;